requirements.txt
beautifulsoup4
requests
psycopg2-binary
pypdf
python-docx
sentence-transformers
faiss-cpu
torch
numpy
tqdm
